requirements.txt
colorama==0.4.4
discord.py==2.0.0
wavelink==1.3.3
